requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
langchain>=0.1.0
langchain-openai>=0.0.5
langgraph>=0.0.20
supabase>=2.3.0
python-dotenv>=1.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
httpx>=0.26.0
pillow>=10.0.0
tavily-python>=0.3.0
openai>=1.10.0
apscheduler>=3.10.0
